% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mediaconnect_operations.R
\name{delete_flow}
\alias{delete_flow}
\title{Deletes a flow}
\usage{
delete_flow(FlowArn)
}
\arguments{
\item{FlowArn}{[required] The ARN of the flow that you want to delete.}
}
\description{
Deletes a flow. Before you can delete a flow, you must stop the flow.
}
\section{Accepted Parameters}{
\preformatted{delete_flow(
  FlowArn = "string"
)
}
}

