% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mediaconnect_operations.R
\name{describe_flow}
\alias{describe_flow}
\title{Displays the details of a flow}
\usage{
describe_flow(FlowArn)
}
\arguments{
\item{FlowArn}{[required] The ARN of the flow that you want to describe.}
}
\description{
Displays the details of a flow. The response includes the flow ARN, name, and Availability Zone, as well as details about the source, outputs, and entitlements.
}
\section{Accepted Parameters}{
\preformatted{describe_flow(
  FlowArn = "string"
)
}
}

