% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mediaconnect_operations.R
\name{grant_flow_entitlements}
\alias{grant_flow_entitlements}
\title{Grants entitlements to an existing flow}
\usage{
grant_flow_entitlements(Entitlements, FlowArn)
}
\arguments{
\item{Entitlements}{[required] The list of entitlements that you want to grant.}

\item{FlowArn}{[required] The flow that you want to grant entitlements on.}
}
\description{
Grants entitlements to an existing flow.
}
\section{Accepted Parameters}{
\preformatted{grant_flow_entitlements(
  Entitlements = list(
    list(
      Description = "string",
      Encryption = list(
        Algorithm = "aes128"|"aes192"|"aes256",
        KeyType = "static-key",
        RoleArn = "string",
        SecretArn = "string"
      ),
      Name = "string",
      Subscribers = list(
        "string"
      )
    )
  ),
  FlowArn = "string"
)
}
}

