% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mediaconnect_operations.R
\name{list_tags_for_resource}
\alias{list_tags_for_resource}
\title{Lists all tags associated with the resource}
\usage{
list_tags_for_resource(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource that you want to view tags for.}
}
\description{
Lists all tags associated with the resource.
}
\section{Accepted Parameters}{
\preformatted{list_tags_for_resource(
  ResourceArn = "string"
)
}
}

