% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mediaconnect_package.R
\docType{package}
\name{paws.mediaconnect-package}
\alias{paws.mediaconnect}
\alias{paws.mediaconnect-package}
\title{paws.mediaconnect: AWS MediaConnect}
\description{
API for AWS Elemental MediaConnect
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
