% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mediaconnect_operations.R
\name{remove_flow_output}
\alias{remove_flow_output}
\title{Removes an output from an existing flow}
\usage{
remove_flow_output(FlowArn, OutputArn)
}
\arguments{
\item{FlowArn}{[required] The flow that you want to remove an output from.}

\item{OutputArn}{[required] The ARN of the output that you want to remove.}
}
\description{
Removes an output from an existing flow. This request can be made only on an output that does not have an entitlement associated with it. If the output has an entitlement, you must revoke the entitlement instead. When an entitlement is revoked from a flow, the service automatically removes the associated output.
}
\section{Accepted Parameters}{
\preformatted{remove_flow_output(
  FlowArn = "string",
  OutputArn = "string"
)
}
}

