% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mediaconnect_operations.R
\name{revoke_flow_entitlement}
\alias{revoke_flow_entitlement}
\title{Revokes an entitlement from a flow}
\usage{
revoke_flow_entitlement(
  EntitlementArn = "string",
  FlowArn = "string"
)
}
\arguments{
\item{EntitlementArn}{[required] The ARN of the entitlement that you want to revoke.}

\item{FlowArn}{[required] The flow that you want to revoke an entitlement from.}
}
\description{
Revokes an entitlement from a flow. Once an entitlement is revoked, the content becomes unavailable to the subscriber and the associated output is removed.
}
\examples{

}
