% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mediaconnect_operations.R
\name{update_flow_entitlement}
\alias{update_flow_entitlement}
\title{You can change an entitlement's description, subscribers, and encryption}
\usage{
update_flow_entitlement(
  Description = "string",
  Encryption = list(
    Algorithm = "aes128"|"aes192"|"aes256",
    KeyType = "static-key",
    RoleArn = "string",
    SecretArn = "string"
  ),
  EntitlementArn = "string",
  FlowArn = "string",
  Subscribers = list(
    "string"
  )
)
}
\arguments{
\item{Description}{A description of the entitlement. This description appears only on the AWS Elemental MediaConnect console and will not be seen by the subscriber or end user.}

\item{Encryption}{The type of encryption that will be used on the output associated with this entitlement.}

\item{EntitlementArn}{[required] The ARN of the entitlement that you want to update.}

\item{FlowArn}{[required] The flow that is associated with the entitlement that you want to update.}

\item{Subscribers}{The AWS account IDs that you want to share your content with. The receiving accounts (subscribers) will be allowed to create their own flow using your content as the source.}
}
\description{
You can change an entitlement's description, subscribers, and encryption. If you change the subscribers, the service will remove the outputs that are are used by the subscribers that are removed.
}
\examples{

}
